+++ core/src/com/giancola/battleship/actors/PlayerBoard.kt
@@ -1,16 +1,20 @@
-package com.giancola.battleship
+package com.giancola.battleship.actors
 
-import com.badlogic.gdx.scenes.scene2d.Actor
 import com.badlogic.gdx.scenes.scene2d.Group
+import com.giancola.battleship.GameConstants
 import com.giancola.battleship.GameConstants.TILE_SIZE
+import com.giancola.battleship.LayoutConstants
 import com.giancola.battleship.screens.PlacementScreen
 
 
-class PlayerBoard(val screen: PlacementScreen, val nRows: Int = GameConstants.N_ROWS, val nCols: Int = GameConstants.N_COLS): Group() {
+class PlayerBoard(val screen: PlacementScreen, val nRows: Int = GameConstants.N_ROWS, val nCols: Int = GameConstants.N_COLS): Group(), DropTarget {
+    override var targetable = true
+
     val tiles = Array(nRows) { i -> Array(nCols) { j ->
         Tile(this, i, j).also {
             it.setPosition(TILE_SIZE * i, TILE_SIZE * j)
             it.setSize(TILE_SIZE, TILE_SIZE)
+
             this.addActor(it)
         }
     } }
@@ -18,5 +22,6 @@ class PlayerBoard(val screen: PlacementScreen, val nRows: Int = GameConstants.N_
     init {
         val boardLayout = LayoutConstants.standard2worldCoords(LayoutConstants.playerBoard)
         this.setPosition(boardLayout.x, boardLayout.y)
+        this.setSize(TILE_SIZE * nCols, TILE_SIZE * nRows)
     }
 }