+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 
+## [0.2] - 2022-02-18
+### Added
+- ship placement ui (drag and drop, rotation)
+- placement validation
+- confirm button
+
+
 ## [0.1] - 2022-02-12
 
 ### Added